users/requirements.txt
Flask==3.0.0
Flask-Cors
psycopg2-binary